Factor out error codes to their respective script modules in order to favor consistency
