+++ hydra-node/test/Hydra/Chain/Direct/Contract/Contest.hs
@@ -12,10 +12,8 @@ import Data.Maybe (fromJust)
 import Cardano.Api.UTxO as UTxO
 import Hydra.Chain.Direct.Contract.Gen (genForParty, genHash, genMintedOrBurnedValue)
 import Hydra.Chain.Direct.Contract.Mutation (
-  HeadError (ChangedParameters, ContesterNotIncluded, HeadValueIsNotPreserved, MustNotPushDeadline, MustPushDeadline, SignerAlreadyContested),
   Mutation (..),
   SomeMutation (..),
-  UtilError (MintingOrBurningIsForbidden),
   addParticipationTokens,
   changeHeadOutputDatum,
   changeMintedTokens,
@@ -26,13 +24,23 @@ import Hydra.Chain.Direct.Contract.Mutation (
   replacePolicyIdWith,
   replaceSnapshotNumber,
   replaceUtxoHash,
-  toErrorCode,
  )
 import Hydra.Chain.Direct.Fixture (testNetworkId, testPolicyId)
 import Hydra.Chain.Direct.Tx (ClosedThreadOutput (..), contestTx, mkHeadId, mkHeadOutput)
 import Hydra.ContestationPeriod (ContestationPeriod, fromChain)
+import Hydra.Contract.Head (
+  HeadError (
+    ChangedParameters,
+    ContesterNotIncluded,
+    HeadValueIsNotPreserved,
+    MustNotPushDeadline,
+    MustPushDeadline,
+    SignerAlreadyContested
+  ),
+ )
 import qualified Hydra.Contract.HeadState as Head
 import Hydra.Contract.HeadTokens (headPolicyId)
+import Hydra.Contract.Util (UtilError (MintingOrBurningIsForbidden), toErrorCode)
 import Hydra.Crypto (HydraKey, MultiSignature, aggregate, sign, toPlutusSignatures)
 import Hydra.Data.ContestationPeriod (posixFromUTCTime)
 import qualified Hydra.Data.ContestationPeriod as OnChain
